requirements.txt
discord.py==2.5.2
google-generativeai==0.7.2
chromadb==1.0.15
python-dotenv==1.1.1
duckduckgo-search==7.5.5
